SONARSLANG-673 Check file status against cached hash
Fix a bug where during PR analysis, a file that had been changed in the
base branch but not the PR (ie: PR branch is out of sync with base
branch), would be incorrectly identified as unchanged by SLang
analyzers.

To ensure that we do not try to reuse cached results for files that have
changed, we rely on the MD5 hash of the file and check it against the
hash in the cache.